# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  aslon1213_gift_pkg_repository.Empty:
    type: object
...
  aslon1213_gift_pkg_repository.Group:
    properties:
      created_at:
        type: string
      id:
        type: string
      member_ids:
        items:
          type: string
        type: array
      name:
        type: string
      owner_id:
        type: string
      updated_at:
        type: string
    type: object
...
  aslon1213_gift_pkg_repository.Response-array_aslon1213_gift_pkg_repository_Group:
    properties:
      data:
        items:
          $ref: '#/definitions/aslon1213_gift_pkg_repository.Group'
        type: array
      message:
        example: ok
        type: string
      status:
        example: success
        type: string
    type: object
...
  aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty:
    properties:
      data:
        $ref: '#/definitions/aslon1213_gift_pkg_repository.Empty'
      message:
        example: ok
        type: string
      status:
        example: success
        type: string
    type: object
...
  aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Group:
    properties:
      data:
        $ref: '#/definitions/aslon1213_gift_pkg_repository.Group'
      message:
        example: ok
        type: string
      status:
        example: success
        type: string
    type: object
...
  pkg_handlers.CreateGroupRequest:
    properties:
      member_ids:
        example:
        - '["60d5ec49f99a2d3a829a7d1f"]'
        items:
          type: string
        type: array
      name:
        example: Book Club
        type: string
      owner_id:
        example: 60d5ec49f99a2d3a829a7d1e
        type: string
    type: object
...
  pkg_handlers.InviteMemberRequest:
    properties:
      member_id:
        example: 60d5ec49f99a2d3a829a7d1e
        type: string
    type: object
...
  pkg_handlers.RemoveMemberRequest:
    properties:
      member_id:
        example: 60d5ec49f99a2d3a829a7d1e
        type: string
    type: object
...
  pkg_handlers.UpdateGroupRequest:
    properties:
      member_ids:
        example:
        - '["60d5ec49f99a2d3a829a7d1f"]'
        items:
          type: string
        type: array
      name:
        example: Book Club
        type: string
      owner_id:
        example: 60d5ec49f99a2d3a829a7d1e
        type: string
    type: object
...
host: localhost:3000
info:
  contact:
    email: hamidovaslon1@gmail.com
    name: API Support
    url: https://github.com/aslon1213/gift
  description: This is a Gift API server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com/aslon1213/gift
  title: Gift API
  version: 0.1.0
paths:
...
  /groups:
    get:
      description: Query all groups for the user by name. Returns groups where the
        user is the owner or a member.
      parameters:
      - description: Group name
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-array_aslon1213_gift_pkg_repository_Group'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
      security:
      - BearerAuth: []
      summary: List or filter groups for the user
      tags:
      - groups
    post:
      consumes:
      - application/json
      description: Create a new group
      parameters:
      - description: Group create request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/pkg_handlers.CreateGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Group'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
      security:
      - BearerAuth: []
      summary: Create group
      tags:
      - groups
  /groups/{id}:
    delete:
      description: Delete a group by id
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
      security:
      - BearerAuth: []
      summary: Delete group
      tags:
      - groups
    get:
      description: Get a specific group by its ID
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Group'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
      security:
      - BearerAuth: []
      summary: Get group by ID
      tags:
      - groups
    put:
      consumes:
      - application/json
      description: Update a group's properties by id
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Group update request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/pkg_handlers.UpdateGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Group'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
      security:
      - BearerAuth: []
      summary: Update group
      tags:
      - groups
  /groups/{id}/invite:
    post:
      consumes:
      - application/json
      description: Invite a member to a group (only owner can invite)
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Invite member request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/pkg_handlers.InviteMemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Group'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
      security:
      - BearerAuth: []
      summary: Invite member to group
      tags:
      - groups
  /groups/{id}/remove:
    post:
      consumes:
      - application/json
      description: Remove a member from a group (only owner can remove)
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Remove member request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/pkg_handlers.RemoveMemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Group'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
      security:
      - BearerAuth: []
      summary: Remove member from group
      tags:
      - groups
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"